Moving `valid()` middleware out of the factory
